requirements.txt
fastapi==0.115.12
pydantic==2.11.2
python-dotenv==1.1.0
uvicorn==0.34.0
langchain-google-alloydb-pg>=0.7.0
google-genai==1.9.0
google-cloud-aiplatform[agent_engines,langchain]
langchain-google-vertexai
langchain-community
pip
ipykernel
pg8000
